requirements.txt
asgiref==3.6.0
Django==3.2
django-ckeditor==6.5.1
django-cors-headers==4.2.0
django-filter==23.2
django-js-asset==2.0.0
djangorestframework==3.14.0
fontawesome-free==5.15.4
Pillow==9.5.0
pytz==2023.3
sqlparse==0.4.4
whitenoise==6.5.0
